Allow for path and filename discrepencies
Change in getDateFromTitle allows files that do not conform to the expected date format (but are still mkvs) to be added with the default date os 1970-01-01.

Change at line 342 makes the channel match logic simply check if the channel name/ID exists in the plex-provided path, rather than trying to match the whole path exactly. This is helpful if the path plex sees is not identical to the filepath that the script sees, ex: if Plex is run in a docker, and the script is run outside the docker.